+++ NEWS.md
@@ -10,6 +10,9 @@
 
 * Wrote `padjust()`, an S3 wrapper for `p.adjust()`, which can also adjust `tableby()` (and hence `paired()`) objects.
 
+* `print.summary.tableby()`, `as.data.frame.summary.tableby()`, `print.summary.modelsum()`, `as.data.frame.summary.modelsum()`: added `width=` and
+  `min.split=` as formal arguments.
+
 # arsenal v1.4.0
 
 **There are a few non-backwards-compatible updates.**